Image editor: slim down to image-loading; park legacy UV editor
The image editor fused two concerns: loading/displaying images and a full
UV editor (UVEditor component, UV picking, select/transform/flag ToolOps,
unwrap tools sidebar). To support sculptcore and a future UV-editing
abstraction, split them apart.

- Park the legacy implementation, unwired, under
  scripts/editors/image/pending-port/ (ImageEditor.ts, uv_ops.js,
  uv_selectops.js, uv_transformops.js) with a TODO.md port checklist.
  Nothing imports them, so they fall out of the TS module graph (tsconfig
  uses `files`) and are neither typechecked nor bundled. ImageBus.ts stays
  put: it is a cross-cutting bus token still used by the mesh unwrapper.

- New minimal ImageEditor: loads images into ImageBlocks (image.open(),
  Ctrl+O, DataBlockBrowser) and displays the active image with a checker
  background + textured quad and pan/zoom. Owns the ImageUser directly, so
  the active image now lives at `imageEditor.imageUser.image` (was
  `imageEditor.uvEditor.imageUser.image`).

- Update cross-layer consumers to the new path / decouple them:
  context.activeTexture, image_ops.js dataPath default, and
  mesh_uvops_base.ts (stops reaching into editor.uvEditor; defaults
  selectedFacesOnly=true until the new UV layer restores a binding).
  Regenerate the datapath catalog and update documentation/editors.md.

A new UV-editing abstraction layer will be designed in a follow-up plan;
see scripts/editors/image/pending-port/TODO.md.
